+++ frontend/src/types/profile/profile.dto.ts
@@ -4,11 +4,11 @@ export interface ProfileDto {
     username: string;
     fullname: string;
     password: string;
-    "activationLink": string;
-    "activated": true;
-    "banned": true;
-    "banReason": string;
-    "status": string;
-    "about": string;
-    "avatar": string;
+    activationLink: string;
+    activated: true;
+    banned: true;
+    banReason: string;
+    status: string;
+    about: string;
+    avatar: string;
 }